requirements.txt
streamlit>=1.30.0
faiss-cpu>=1.7.4
sentence-transformers>=2.3.1
groq>=0.4.1
PyPDF2>=3.0.0
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.1
